backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.23
asyncpg>=0.27.0        # Async PostgreSQL driver
psycopg2-binary>=2.9.1
pydantic>=1.8.2
python-multipart>=0.0.5
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
requests>=2.26.0
aiohttp>=3.8.1         # Async HTTP client
aiofiles>=0.8.0        # Async file operations
pillow>=8.3.1
numpy>=1.21.2
geopandas>=0.12.0
pandas>=1.5.0
matplotlib>=3.6.0      # For geopandas visualization
pysolar>=0.10.0        # For sunlight calculations
rasterio>=1.3.0        # For satellite imagery processing
geopy>=2.3.0           # For geocoding addresses
pyproj>=3.4.0          # For coordinate system transformations
scikit-image>=0.19.0   # For image processing
scipy>=1.9.0           # For scientific calculations
python-dotenv>=0.19.0  # For environment variables
pytest>=6.2.5          # For testing
pytest-asyncio>=0.18.0 # For async tests
httpx>=0.23.0          # For async HTTP testing
